+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Improvements:
  query parameters
 * [OE-86](https://openlmis.atlassian.net/browse/OE-86): GET /programs has new program code repeatable query parameter
 * [OLMIS-8099](https://openlmis.atlassian.net/browse/OLMIS-8099): Delete unsed report rights, create new superset report rights
+* [OE-87](https://openlmis.atlassian.net/browse/OE-87): GET /orderables and /lots have new repeatebale `exectCode` query
+ parameter to filter by exact codes (opposed to `code` doing contains condition)
 
 Bux fixes:
 * [OLMIS-8068](https://openlmis.atlassian.net/browse/OLMIS-8068): Fix NPE on adding programs to products